refactor: clean up unused using directives and simplify property initializations
