+++ Domain/Fakes/FakeAppUserDto.cs
@@ -7,9 +7,6 @@
 // Project Name :  Domain
 // =======================================================
 
-using Domain.Enums;
-using Domain.Models;
-
 namespace Domain.Fakes;
 
 /// <summary>
