+++ Domain/Fakes/FakeArticle.cs
@@ -7,8 +7,6 @@
 // Project Name :  Domain
 // =======================================================
 
-using Domain.Entities;
-
 namespace Domain.Fakes;
 
 /// <summary>
