+++ Domain/Fakes/FakeCategory.cs
@@ -7,9 +7,6 @@
 // Project Name :  Domain
 // =======================================================
 
-using Domain.Entities;
-using Domain.Enums;
-
 namespace Domain.Fakes;
 
 /// <summary>
@@ -55,7 +52,7 @@ public class FakeCategory
 		var fake = new Faker<Category>()
 				.RuleFor(x => x.Id, ObjectId.GenerateNewId())
 				.RuleFor(x => x.Name, f => f.Random.Enum<CategoryNames>().ToString())
-				.RuleFor(x => x.Description, (f, x) => $"This category is for {x.Name} related items.");
+				.RuleFor(x => x.Description, (_, x) => $"This category is for {x.Name} related items.");
 
 		return useSeed ? fake.UseSeed(seed) : fake;
 	}
