+++ Domain/GlobalUsings.cs
@@ -10,8 +10,6 @@
 global using System;
 global using System.Collections.Generic;
 global using System.ComponentModel.DataAnnotations;
-global using System.Linq;
-global using System.Linq.Expressions;
 global using System.Threading;
 global using System.Threading.Tasks;
 global using System.Web;
