+++ Domain/Interfaces/ICategoryRepository.cs
@@ -7,8 +7,6 @@
 // Project Name :  Domain
 // =======================================================
 
-using Domain.Entities;
-
 namespace Domain.Interfaces;
 
 public interface ICategoryRepository : IRepository<Category> { }
