+++ Domain/Models/AppUserDto.cs
@@ -44,7 +44,7 @@ public class AppUserDto
 	/// <summary>
 	///  Gets or sets the unique identifier for the user.
 	/// </summary>
-	public string Id { get; set; } = string.Empty;
+	public string Id { get; set; }
 
 	/// <summary>
 	///  Gets or sets the username of the user.
