+++ Persistence.MongoDb/GlobalUsings.cs
@@ -7,7 +7,6 @@
 global using System;
 global using System.Collections.Generic;
 global using System.Linq;
-global using System.Linq.Expressions;
 global using System.Threading.Tasks;
 
 global using Mapster;
