+++ Tests/Architecture.Tests/GlobalUsings.cs
@@ -5,7 +5,6 @@
 // =======================================================
 
 global using System;
-global using System.Collections;
 global using System.Collections.Generic;
 global using System.Diagnostics.CodeAnalysis;
 global using System.Linq;
@@ -13,15 +12,8 @@ global using System.Reflection;
 
 global using FluentAssertions;
 
-global using Microsoft.Extensions.DependencyInjection;
-
-global using MyMediator;
-
 global using NetArchTest.Rules;
 
 global using Domain.Abstractions;
-global using Domain.Interfaces;
-
-global using Web;
 
 global using Xunit;
