+++ Web/Components/Features/Categories/Components/CategoryDetails.razor
@@ -1,7 +1,6 @@
 @page "/categories/details/{Id}"
 
 @inject ICategoryService CategoryService
-@inject NavigationManager NavigationManager
 
 <div class="container mx-auto px-4 py-8">
 	<div class="mb-8">
