+++ Web/Components/Shared/PostInfoComponent.razor
@@ -1,4 +1,3 @@
-@using System.Globalization
 <div class="flex gap-4 border-t border-gray-200 pt-4">
 
 	<div>Author: @Article.Author.UserName</div>
