+++ Web/GlobalUsings.cs
@@ -8,16 +8,10 @@
 // =======================================================
 
 global using System;
-global using System.Collections.Generic;
-global using System.Linq;
-global using System.Threading;
-global using System.Threading.Tasks;
 
 global using Microsoft.Extensions.DependencyInjection;
 
 global using MongoDB.Bson;
 
-global using Domain.Abstractions;
-global using Domain.Entities;
 global using Domain.Models;
 global using Web.Components;
