+++ Web/Program.cs
@@ -1,3 +1,5 @@
+using ServiceDefaults;
+
 var builder = WebApplication.CreateBuilder(args);
 
 // Add service defaults & Aspire client integrations.
